+++ include/readfile.h
@@ -1,0 +1,7 @@
+#ifndef READFILE_H
+#define READFILE_H
+
+char *readFile(const char *filename);
+void freeFile(char *fileContent);
+
+#endif //READFILE_H